Optimal Scheduling (#10)
Completed first implementation of optimal scheduling functionality. This uses a DFS branch and bound search of the solution space to find a solution with the lowest possible schedule length. 

This has been tested on all provided inputs and produces the correct output on both 2 and 4 processors.

Further testing is required and optimizations can still be made.